OSM: add cast to avoid potential undefined behaviour with the default signed int type
git-svn-id: https://svn.osgeo.org/gdal/trunk@39669 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/osm/osm_parser.cpp
@@ -1873,7 +1873,8 @@ static OSMRetCode PBF_ProcessBlock(OSMContext* psCtxt)
             break;
         }
         const unsigned int nHeaderSize =
-            (abyHeaderSize[0] << 24) | (abyHeaderSize[1] << 16) |
+            (static_cast<unsigned int>(abyHeaderSize[0]) << 24) |
+            (abyHeaderSize[1] << 16) |
             (abyHeaderSize[2] << 8) | abyHeaderSize[3];
 
         psCtxt->nBytesRead += 4;
